Move status reply into a seperate file
